requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
GoogleBard==1.2.2
h11==0.14.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.2
pandas-stubs==1.5.2.221124
prompt-toolkit==3.0.38
pycparser==2.21
Pygments==2.15.1
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==13.4.2
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
style==1.1.0
telegram==0.0.1
tornado==6.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
types-pytz==2022.6.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
update==0.0.1
urllib3==1.26.13
wcwidth==0.2.6
wsproto==1.2.0
yarl==1.9.2
